+++ lib/constants.py
@@ -1,6 +1,6 @@
 PAGE_TITLE_PREFIX = "Dash Pip Components | "
 PRIMARY_COLOR = "teal"
-APP_VERSION = "0.3.0"
+APP_VERSION = "1.0.0"
 
 # This will be populated by pages/markdown.py when loading documentation files
 NAME_CONTENT_MAP = {}